Update to latest operator-rs main branch and remove debug kuttl steps.
