+++ tests/kuttl-test.yaml.jinja2
@@ -8,3 +8,4 @@ testDirs:
 
 startKIND: false
 suppress: ["events"]
+parallel: 2
