+++ tests/templates/kuttl/spark-pi-private-s3/10-assert.yaml
@@ -2,13 +2,13 @@
 apiVersion: kuttl.dev/v1beta1
 kind: TestAssert
 metadata:
-  name: spark-pi
+  name: spark-pi-private-s3
 timeout: 900
 ---
 # The Job starting the whole process
 apiVersion: batch/v1
 kind: Job
 metadata:
-  name: spark-pi
+  name: spark-pi-private-s3
 status:
   succeeded: 1
