--- tests/templates/kuttl/spark-pi-public-s3/03-assert.yaml
@@ -1,7 +1,0 @@
----
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: list-jars-job
-status:
-  succeeded: 1
